Flatten our Dev14/Dev15 completion object hierarchy
Previously, we supported a light-up ability where we would use the
appropriate completion APIs for Dev14 or Dev15 depending upon which
version of Visual Studio we were targeting. We no longer support such
a configuration, so we can flatten all the code and indirection back
down.
