+++ src/Interactive/EditorFeatures/Core/Implementation/Completion/Presentation/CompletionPresenter.cs
@@ -7,11 +7,7 @@ using Microsoft.VisualStudio.Language.Intellisense;
 using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Text.Editor;
 using Microsoft.VisualStudio.Utilities;
-using Microsoft.VisualStudio.InteractiveWindow;
 using Microsoft.VisualStudio.InteractiveWindow.Commands;
-using System.Collections.Generic;
-using System;
-using Microsoft.CodeAnalysis.Editor.Extensibility.Composition;
 
 namespace Microsoft.CodeAnalysis.Editor.Implementation.Completion.Presentation
 {
@@ -22,24 +18,21 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.Completion.Presentation
     {
         private readonly ICompletionBroker _completionBroker;
         private readonly IGlyphService _glyphService;
-        private readonly ICompletionSetFactory _completionSetFactory;
 
         [ImportingConstructor]
         public CompletionPresenter(
             ICompletionBroker completionBroker,
-            IGlyphService glyphService,
-            [ImportMany] IEnumerable<Lazy<ICompletionSetFactory, VisualStudioVersionMetadata>> completionSetFactories)
+            IGlyphService glyphService)
         {
             _completionBroker = completionBroker;
             _glyphService = glyphService;
-            _completionSetFactory = VersionSelector.SelectHighest(completionSetFactories);
         }
 
         ICompletionPresenterSession IIntelliSensePresenter<ICompletionPresenterSession, ICompletionSession>.CreateSession(ITextView textView, ITextBuffer subjectBuffer, ICompletionSession sessionOpt)
         {
             AssertIsForeground();
             return new CompletionPresenterSession(
-                _completionSetFactory, _completionBroker, _glyphService, textView, subjectBuffer);
+                _completionBroker, _glyphService, textView, subjectBuffer);
         }
 
         ICompletionSource ICompletionSourceProvider.TryCreateCompletionSource(ITextBuffer textBuffer)
